Cluster Fuck change according to Dream

+++ Source/Runtime/Win32/Win32Error.cpp
@@ -18,7 +18,7 @@ namespace Portakal
         /**
          * Get message string from win32
          */
-        const unsigned long long sizeInBytes = FormatMessageA(FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS,
+        const UInt64 sizeInBytes = FormatMessageA(FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS,
             NULL, errorMessageID, MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), (LPSTR)&messageBuffer, 0, NULL);
 
         String message(messageBuffer, sizeInBytes);
